--- database/elastic_mapping/text_sorting.js
@@ -1,11 +1,0 @@
-const { USE_ELASTIC_ICU } = process.env;
-
-const textSorting = {};
-
-if (USE_ELASTIC_ICU === 'true') {
-  [textSorting.type, textSorting.numeric] = ['icu_collation_keyword', true];
-} else {
-  [textSorting.type, textSorting.fielddata, textSorting.analyzer] = ['text', true, 'string_sorter'];
-}
-
-export default textSorting;
